Move division classification into track_errors and start to eliminate use of DivisionEvents
